Remove custom signature file before processing it
The Google Exposure Notification API won't accept zip files with additional files in it, like the added signature. While processing the file, stream the zip and write only the valid entries to the input zip file, whilst extracting the signature in memory so that it can be validated against the content.bin afterwards.

Signature validation is still todo, but this allows for processing the files through the EN api.