Add reminder notification for exposures
When the app is not opened after an exposure notification is shown, remind the user every 3 hours.
ExposureNotificationJob will queue this reminder job, and starting the app or using the action in the reminder notification will cancel it.

Update the notification channels to vibrate for exposure notifications, and set the default privacy level to secret on that channel too (although this doesn't always seem to work).