Clean up handling of response signatures and accept headers
Remove the use of accept headers for development as this isn't working anymore anyway. SignedResponseInterceptor can be simplified for that reason.
The feature flag now turns validation of the responses on or off, independent of the accept header setting.
Let exposure key file response signature check depend on the api build config, since they are essentially the same thing.