Workaround JobScheduler dropping jobs on some platforms
WorkManager will not reschedule jobs on job scheduler if they have been removed for any reason. On some versions of Android a job could be dropped if it can't be executed. As a result this would go unnoticed and the job would only be executed when the app is in de foreground.

To workaround a "forced stopped" situation is emulated, which wil make WorkManager always reschedule jobs when it's initialised. An upcoming version of WorkManager will have a better fix for this, so this is a temporary measure.

As a second workaround a periodic alarm is scheduled to wake up or schedule a job. This will reduce the change of a job failing and trigger the rescheduling if the app process was no longer alive or the execution of overdue jobs if possible.