Disable EN when the end of life screen is shown
When the app is end of life, all processing and Exposure Notifications will be disabled, however since this happens as part of background work, these tasks might not been disabled yet when the screen is shown.

To make sure the expected state is correct, disable all our processing in the same way when the end of life screen is presented to the user.