Retry creation of EncryptedShared preferences
Creating EncryptedSharedPreferences might fail due to contention on the underlying key store. As per recommendation in https://issuetracker.google.com/issues/158234058 creation of EncryptedSharedPreferences is now retried with a backoff. When retries are exhausted, the exception thrown is wrapped in a RetryException.

Added a loading spinner to the status screen, that shows until the initial status is emitted. Since this relies on the preferences being initialised this effectively means that the status screen will show a spinner until the preferences are initialised.

All operations on the encrypted preferences are now async, through the AsyncSharedPreferences wrapper class. Fixes Github issue #32