Update WorkManager to 2.5.0-beta02
Update WorkManager and remove the previous workaround for forcing jobs to be rescheduled when they are lost on Huawei devices.
Remove the PeriodicWakeUpReceiver as Play Services is waking the app more often now.
Add UpdateReceiver to get an opportunity to get woken up and reschedule jobs when an update is installed.