Fix dashboard transitions
Separate 3 cases for transition on DashboardFragment

* Deeplinking into a dashboard (graph)
* From dashboard to another one (replacing the screen)
* From the dashboard overview

When jumping into a graph from the StatusFragment, trampoline through the DashboardOverviewFragment. This seems to be needed to get the shared element (appbar) transition working correctly.

Navigation graph has been split so that the view model can be shared between the dashboard destinations without reloading the dashboard data every time (from network or cache).